Refactor DarkModeService and update modal content
Reorganized DarkModeService into a theme directory, updating import paths in `theme-toggler.component.ts`, `base-layout.component.ts`, and `app.component.ts`. Improved modal title and body text in `weather-forecast-page.component.html` for better clarity. Reformatted `dark-mode.service.spec.ts` and `dark-mode.service.ts` files. Updated hash values in `_metadata.json`.
