+++ src/DevServerClient/.angular/cache/19.1.6/DevServerClient/vite/deps_ssr/_metadata.json
@@ -1,8 +1,8 @@
 {
-  "hash": "94071579",
+  "hash": "20d220ca",
   "configHash": "612d7e25",
-  "lockfileHash": "9168342d",
-  "browserHash": "b64805e2",
+  "lockfileHash": "5e1e97cd",
+  "browserHash": "2341a1fa",
   "optimized": {},
   "chunks": {}
 }
