+++ src/DevServerClient/src/app/services/theme/dark-mode.service.spec.ts
@@ -1,5 +1,4 @@
 import { TestBed } from '@angular/core/testing';
-
 import { DarkModeService } from './dark-mode.service';
 
 describe('DarkModeService', () => {
